# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple example NodeJS API project to demonstrate Swagger Documentation",
        "version": "1.0.0",
        "title": "Tasks API",
        "contact": {
            "email": "abc@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "schemes": [
        "http"
    ],
    "host": "localhost:3000",
    "basePath": "/auth",
    "paths": {
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all the users",
                "description": "Get all the users",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array"
                        }
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get one user",
                "description": "Get one user",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id that needs to be found",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "update user",
                "description": "update user",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id for the targeted user",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "in": "body",
                        "description": "update user",
                        "name": "updated information",
                        "required": true,
                        "type": "object",
                        "schema": {
                            "properties": {
                                "fullName": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string",
                                    "format": "email"
                                },
                                "phone_number": {
                                    "type": "integer"
                                },
                                "password": {
                                    "type": "string",
                                    "format": "password"
                                },
                                "role_id": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful update",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "delete user",
                "description": "delete user",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "delete user",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "deleted user",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/signup": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Add new user",
                "description": "Add new user",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "new user",
                        "required": true,
                        "type": "object",
                        "schema": {
                            "properties": {
                                "fullName": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string",
                                    "format": "email"
                                },
                                "phone_number": {
                                    "type": "integer"
                                },
                                "password": {
                                    "type": "string",
                                    "format": "password"
                                },
                                "role_id": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "user login",
                "description": "user login",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "user login",
                        "required": true,
                        "type": "object",
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "format": "email"
                                },
                                "password": {
                                    "type": "string",
                                    "format": "password"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful login",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid user name or password"
                    }
                }
            }
        },
...
    }
}